+++ src/BaGet.Web/Controllers/CacheIndexModule.cs
@@ -8,6 +8,10 @@ using Newtonsoft.Json;
 using Carter.ModelBinding;
 using Carter.Request;
 using Carter.Response;
+using BaGet.Core.Configuration;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Routing;
+using System.Threading.Tasks;
 
 namespace BaGet.Web.Controllers
 {
@@ -27,8 +31,8 @@ namespace BaGet.Web.Controllers
             }
         }
 
-        public CacheIndexModule() {
-            this.Get(prefix + "/v3/index.json", async (req, res, routeData) =>
+        public CacheIndexModule(LiGetCompatibilityOptions compat) {
+            Func<HttpRequest, HttpResponse, RouteData, Task> indexHandler = async (req, res, routeData) =>           
             {
                 await res.AsJson(new
                 {
@@ -41,7 +45,11 @@ namespace BaGet.Web.Controllers
                         .Concat(ServiceWithAliases("SearchAutocompleteService", req.PackageAutocomplete(prefix), "", "3.0.0-rc", "3.0.0-beta"))
                         .ToList()
                 });
-            });
+            };
+            this.Get(prefix + "/v3/index.json", indexHandler);
+            if(compat != null && compat.Enabled) {
+                this.Get("/api/" + prefix + "/v3/index.json", indexHandler);
+            }
         }
     }
 }