+++ tasks.sh
@@ -74,11 +74,20 @@ case "${command}" in
     source_imagerc "${image_dir}"  "${imagerc_filename}"
     ide --idefile Idefile.e2e-docker "e2e/stress/run.sh"
     ;;
+  liget_compat_docker)
+    source_imagerc "${image_dir}"  "${imagerc_filename}"
+    rm -rf e2e/liget-compat/data/db/*
+    rm -rf e2e/liget-compat/data/packages/*
+    rm -rf e2e/liget-compat/data/cache/*
+    export LiGetCompat__Enabled=true
+    ide --idefile Idefile.liget-compat "e2e/liget-compat/run.sh"
+    ;;
   all)
     ide "./build.sh --target All"
     ./tasks.sh build_docker
     ./tasks.sh test_docker
-    ./tasks.sh stress_docker
+    ./tasks.sh liget_compat_docker
+    ./tasks.sh stress_docker    
     ;;
   prepare_code_release)
     version=$2