+++ tests/BaGet.Tests/NuGetClientIntegrationTest.cs
@@ -17,6 +17,9 @@ using FluentValidation;
 using NuGet.Packaging.Core;
 using NuGet.Versioning;
 using System.IO;
+using System.Xml.Linq;
+using Microsoft.Extensions.DependencyInjection;
+using BaGet.Core.Services;
 
 namespace BaGet.Tests
 {
@@ -26,10 +29,18 @@ namespace BaGet.Tests
     public class NuGetClientIntegrationTest : IDisposable
     {
         private const string CacheIndex = "cache/v3/index.json";
+        private const string CompatCacheIndex = "api/cache/v3/index.json";
         static readonly string MainIndex = "v3/index.json";
-        public static IEnumerable<object[]> TestCases = new[] {
+        static readonly string V2Index = "v2";
+        static readonly string CompatV2Index = "api/v2";
+        public static IEnumerable<object[]> V3Cases = new[] {
             new object[] { MainIndex },
             new object[] { CacheIndex },
+            new object[] { CompatCacheIndex },
+        };
+        public static IEnumerable<object[]> V2Cases = new[] {
+            new object[] { V2Index },
+            new object[] { CompatV2Index }
         };
         protected readonly ITestOutputHelper Helper;
         private readonly TestServer server;
@@ -85,7 +96,7 @@ namespace BaGet.Tests
         }
 
         [Theory]
-        [MemberData(nameof(TestCases))]
+        [MemberData(nameof(V3Cases))]
         public async Task GetIndexShouldReturn200(string indexEndpoint)
         {
             InitializeClient(indexEndpoint);
@@ -95,7 +106,17 @@ namespace BaGet.Tests
         }
 
         [Theory]
-        [MemberData(nameof(TestCases))]
+        [MemberData(nameof(V2Cases))]
+        public async Task GetV2IndexShouldReturn200(string indexEndpoint)
+        {
+            InitializeClient(indexEndpoint);
+            var index = await _httpClient.GetAsync(indexUrl);
+            Assert.Equal(HttpStatusCode.OK, index.StatusCode);
+            return;
+        }
+
+        [Theory]
+        [MemberData(nameof(V3Cases))]
         public async Task IndexResourceHasManyEntries(string indexEndpoint)
         {
             InitializeClient(indexEndpoint);
@@ -104,7 +125,17 @@ namespace BaGet.Tests
         }
 
         [Theory]
-        [MemberData(nameof(TestCases))]
+        [MemberData(nameof(V2Cases))]
+        public async Task V2IndexResourceReturnsODataServiceDocument(string indexEndpoint)
+        {
+            InitializeClient(indexEndpoint);
+            var indexResource = await _sourceRepository.GetResourceAsync<ODataServiceDocumentResourceV2>();
+            Assert.NotNull(indexResource);
+            Assert.Equal("http://localhost/" + indexEndpoint, indexResource.BaseAddress);
+        }
+
+        [Theory]
+        [MemberData(nameof(V3Cases))]
         public async Task IndexIncludesAtLeastOneSearchQueryEntry(string indexEndpoint)
         {
             InitializeClient(indexEndpoint);
@@ -113,7 +144,7 @@ namespace BaGet.Tests
         }
 
         [Theory]
-        [MemberData(nameof(TestCases))]
+        [MemberData(nameof(V3Cases))]
         public async Task IndexIncludesAtLeastOneRegistrationsBaseEntry(string indexEndpoint)
         {
             InitializeClient(indexEndpoint);
@@ -122,7 +153,7 @@ namespace BaGet.Tests
         }
 
         [Theory]
-        [MemberData(nameof(TestCases))]
+        [MemberData(nameof(V3Cases))]
         public async Task IndexIncludesAtLeastOnePackageBaseAddressEntry(string indexEndpoint)
         {
             InitializeClient(indexEndpoint);
@@ -131,12 +162,36 @@ namespace BaGet.Tests
         }
 
         [Theory]
-        [MemberData(nameof(TestCases))]
+        [MemberData(nameof(V3Cases))]
         public async Task IndexIncludesAtLeastOneSearchAutocompleteServiceEntry(string indexEndpoint)
         {
             InitializeClient(indexEndpoint);
             var indexResource = await _sourceRepository.GetResourceAsync<ServiceIndexResourceV3>();
             Assert.NotEmpty(indexResource.GetServiceEntries("SearchAutocompleteService"));
+        }
+
+        [Fact]
+        [Trait("Category", "integration")] // because it uses external nupkg files
+        public async Task GetV2FindPackagesById()
+        {
+            InitializeClient(MainIndex);
+            var packageResource = await _sourceRepository.GetResourceAsync<PackageUpdateResource>();
+            await packageResource.Push(TestResources.GetNupkgBagetTest1(),
+                null, 5, false, GetApiKey, GetApiKey, false, logger);
+            var response = await _httpClient.GetAsync("/v2/FindPackagesById()?id=baget-test1");
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+            var responseText = await response.Content.ReadAsStringAsync();               
+            var entries = XmlFeedHelper.ParsePage(XDocument.Parse(responseText));
+            var entry = Assert.Single(entries);
+            AssertBaGetTestEntry(entry);
+        }
+
+        private void AssertBaGetTestEntry(V2FeedPackageInfo dummyEntry)
+        {
+            Assert.Equal("baget-test1", dummyEntry.Id);
+            Assert.Equal(NuGetVersion.Parse("1.0.0"), dummyEntry.Version);
+            Assert.Equal("baget-test1", dummyEntry.Authors.Single());
+            Assert.Equal("::netstandard2.0", dummyEntry.Dependencies);
         }
 
         // Push
@@ -153,6 +208,24 @@ namespace BaGet.Tests
             Assert.NotEmpty(meta);
             var one = meta.First();
             Assert.Equal(new PackageIdentity("baget-test1", NuGetVersion.Parse("1.0.0")), one.Identity);
+        }
+
+        [Fact]
+        [Trait("Category", "integration")] // because it uses external nupkg files
+        public async Task PushedPackageShouldExistWithPackageDependenciesInPackageService()
+        {
+            InitializeClient(MainIndex);
+            var packageResource = await _sourceRepository.GetResourceAsync<PackageUpdateResource>();
+            await packageResource.Push(TestResources.GetNupkgBagetTest1(),
+                null, 5, false, GetApiKey, GetApiKey, false, logger);
+
+            var packageService = server.Host.Services.GetRequiredService<IPackageService>();
+            Assert.True(await packageService.ExistsAsync("baget-test1", NuGetVersion.Parse("1.0.0")));
+            var found = await packageService.FindAsync("baget-test1", NuGetVersion.Parse("1.0.0"), false, true);
+            Assert.NotNull(found.Dependencies);
+            Assert.NotEmpty(found.Dependencies);
+            var one = found.Dependencies.Single();
+            Assert.Equal("netstandard2.0", one.TargetFramework);
         }
 
         [Fact]